handle DQT debug form with new mechanism
